Rae language change: `Float` is now f32 (alias of Float32), not f64
Permanent language-design decision. Rae targets games, graphics, UI and
other performance-sensitive native software, where the common representation
for vectors, matrices, transforms, geometry, UI coordinates, colours,
vertices and particles is f32 — and where WGSL's ordinary float is f32.
Defaulting to f64 (the Python/JS convention) cost half the cache density and
bandwidth and imposed a routine f64->f32 narrowing on every GPU upload.
Rationale in full: docs/primitive-types.md.

    Float = Float32 = IEEE-754 binary32   (C float)
    Float64         = IEEE-754 binary64   (C double)

TYPE SYSTEM — Float32/Float64 are now REAL types, not backend-only strings.
Previously the front end had a single TYPE_FLOAT and the C backend decided
width by matching type-name strings; Float64/Float32 were never resolved in
sema at all. Now:
- TYPE_FLOAT is f32; new TYPE_FLOAT64 is a distinct kind (type.h/type.c).
- sema resolves `Float` and `Float32` to the SAME interned TypeInfo, so they
  share identity, equality, ABI, layout and monomorphization: List(Float)
  and List(Float32) are one instantiation by construction.
- `Float64` resolves separately, so List(Float64) is distinct.

MANGLING — fixes a latent collision: `Float` and `Float64` previously
mangled to the same name, which would have given List(Float) and
List(Float64) one monomorphization with DIFFERENT element sizes. Float/
Float32 now mangle as "Float", Float64 as "Float64". type.c had a second,
independent mangler that still emitted "double"; it is now aligned, which
was the cause of rae_List_float being defined but rae_List_double referenced.

C BACKEND — Float/Float32 -> `float`, Float64 -> `double`, emitted as plain
idiomatic C. No Rae-specific numeric typedefs were introduced.

FFI/ABI — the boundary is literal: Buffer(Float) is `float*`, Buffer(Float64)
is `double*`; there is no hidden widening. ~70 extern C signatures were
flipped in lockstep with their Rae declarations. The dangerous case (a Rae
float buffer read by C as double*) was verified to be caught by the C
compiler for header-declared externs; runtime/gpu externs not in the shared
header were audited by hand and by compiling generated C against the runtime.

GPU — removed the mesh-upload temp buffer in runtime_gpu3d.c that existed
solely to narrow f64->f32; vertex data now uploads directly. The i64->u32
index conversion is still semantically required and stays. The frame/draw
uniform repacking also stays: it reorders into padded std140 layout, which
is required independently of width.

INTENTIONAL Float64 — epoch timestamps. `gpu2d.nowSeconds()` and
`sys.fileMtime`/`fileModTime` return Float64 because a ~1.7e9 timestamp in
f32 has ~128s granularity: two calls one frame apart return the SAME value,
so every animation delta would be exactly 0.0 and file-watching would never
observe a change. hot_reload/file_watch `lastMtime` fields and locals were
widened to match; that chain was a real bug this change surfaced (415).

TESTS — new 550_float_is_f32 proves Float has f32 semantics, Float32 is the
same type, Float64 is distinct and retains f64 semantics, the alias is
interchangeable, and List(Float)/List(Float32) coexist. Uses 2^24 (last
integer f32 represents exactly) compared rather than printed, so it does not
depend on float formatting.

DOCS — new docs/primitive-types.md; corrected FFI.md, noise.md and
standard-library-ownership-model.md. noise.md now states that CPU and GPU
share a representation but does NOT claim bit-identical determinism:
operation order, FMA contraction and transcendental precision still differ.

+++ compiler/runtime/runtime_gpu2d_text.c
@@ -194,10 +194,10 @@ static void rae_g2d_ensure_text_inst(int ai, int prims) {
 
 /* Full glyph submit with an optional outline (outlineWidth px + colour) and
  * softness (edge-falloff width in px; 1 = crisp, larger = soft/blurred). */
-void rae_ext_gpu2d_drawGlyphEx(double sx0, double sy0, double sx1, double sy1,
-                               double u0, double v0, double u1, double v1,
-                               int64_t atlas, double pxRange, int64_t color,
-                               double outlineWidth, int64_t outlineColor, double softness) {
+void rae_ext_gpu2d_drawGlyphEx(float sx0, float sy0, float sx1, float sy1,
+                               float u0, float v0, float u1, float v1,
+                               int64_t atlas, float pxRange, int64_t color,
+                               float outlineWidth, int64_t outlineColor, float softness){
     int ai = (int)atlas - 1;
     if (ai < 0 || ai >= RAE_SDF_MAX_ATLAS) return;
     int need = (g_g2d_text_count[ai] + 1) * G2D_TEXT_FLOATS;
@@ -228,9 +228,9 @@ void rae_ext_gpu2d_drawGlyphEx(double sx0, double sy0, double sx1, double sy1,
 }
 
 /* Back-compat: glyph with no outline. */
-void rae_ext_gpu2d_drawGlyph(double sx0, double sy0, double sx1, double sy1,
-                             double u0, double v0, double u1, double v1,
-                             int64_t atlas, double pxRange, int64_t color) {
+void rae_ext_gpu2d_drawGlyph(float sx0, float sy0, float sx1, float sy1,
+                             float u0, float v0, float u1, float v1,
+                             int64_t atlas, float pxRange, int64_t color){
     rae_ext_gpu2d_drawGlyphEx(sx0, sy0, sx1, sy1, u0, v0, u1, v1, atlas, pxRange, color, 0.0, 0, 1.0);
 }
